move classes to packages
