+++ app/src/main/java/net/romanitalian/moneytrackerapp/adapters/TransactionAdapter.java
@@ -1,4 +1,4 @@
-package net.romanitalian.moneytrackerapp;
+package net.romanitalian.moneytrackerapp.adapters;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -6,6 +6,9 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
+
+import net.romanitalian.moneytrackerapp.R;
+import net.romanitalian.moneytrackerapp.models.Transaction;
 
 import java.util.List;
